move dragSource to dragIcon
