+++ src/js/pages/takes/TakeCard/Components/TopBar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-
+import DragSource from './DragTarget';
 export default class TakeCardTopIcon extends React.Component {
 
   constructor(props) {
@@ -12,9 +12,8 @@ export default class TakeCardTopIcon extends React.Component {
 
     return  (
       <TopBar>
-        <DragIcon>
-          <i className = "material-icons">menu</i>
-        </DragIcon>
+
+        <DragSource {...this.props} />
 
         <CardInfo>
           <h3 style={{alignSelf: 'center'}}> {this.props.txt.take} {this.props.take_num} </h3>
